Moved components to their own file
